Core: rework datasource configuration by Spring JPA

+++ Core/src/main/java/com/stanislav/entities/user/User.java
@@ -1,7 +1,6 @@
 package com.stanislav.entities.user;
 
 import jakarta.persistence.*;
-
 import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
@@ -13,7 +12,7 @@ public final class User implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     private String login;
 
@@ -34,11 +33,11 @@ public final class User implements Serializable {
     public User() {}
 
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
